thresholdagent.proto is added and proto structs are used in ceremony

+++ utils/gen_certs.go
@@ -87,9 +87,9 @@ func VerifyCertificate(childCertPem, rootCertPem []byte) bool {
 	return VerifyCert(childCert, rootCert)
 }
 
-func GetId(cert *x509.Certificate) int {
+func GetId(cert *x509.Certificate) int32 {
 	result, _ := strconv.Atoi(string([]byte(cert.Subject.CommonName)[6:]))
-	return result
+	return int32(result)
 }
 func VerifyCert(childCert, rootCert *x509.Certificate) bool {
 	var roots = x509.NewCertPool()
